feat: add carousel plugin for autoplay
