+++ package.json
@@ -20,6 +20,7 @@
     "@tabler/icons-react": "^3.2.0",
     "class-variance-authority": "^0.7.0",
     "clsx": "^2.1.0",
+    "embla-carousel-autoplay": "^8.1.3",
     "embla-carousel-react": "^8.1.3",
     "framer-motion": "^11.2.6",
     "jsdom": "^24.0.0",
